+++ volmdlr/core_compiled.pyx
@@ -3124,17 +3124,22 @@ class Frame2D(Basis2D):
         new_origin = self.origin.translation(vector)
         return Frame2D(new_origin, self.u, self.v)
 
-    def rotation(self, angle):
+    def rotation(self, center: Point2D, angle: float):
         """
         Returns a rotated 2-dimensional frame.
 
+        :param center: The center of rotation
+        :type center: :class:`volmdlr.Point2D`
         :param angle: The rotation angle
         :type angle: float
         :return: New rotated frame
         :rtype: :class:`volmdlr.Frame2D`
         """
+        new_origin = self.origin
+        if not center.is_close(new_origin):
+            new_origin = self.origin.rotation(center, angle)
         new_base = Basis2D.rotation(self, angle)
-        return Frame2D(self.origin, new_base.u, new_base.v)
+        return Frame2D(new_origin, new_base.u, new_base.v)
 
     def Draw(self, ax=None, style="ok"):
         """
@@ -3401,10 +3406,10 @@ class Frame3D(Basis3D):
         :rtype: :class:`volmdlr.Frame3D`
         """
         new_base = Basis3D.rotation(self, axis, angle)
+        if center.is_close(self.origin):
+            return Frame3D(self.origin, new_base.u, new_base.v, new_base.w, self.name)
         new_origin = self.origin.rotation(center, axis, angle)
-        return Frame3D(new_origin,
-                       new_base.u, new_base.v, new_base.w,
-                       self.name)
+        return Frame3D(new_origin, new_base.u, new_base.v, new_base.w, self.name)
 
     def translation(self, offset: Vector3D):
         """